feat(Student): add functionality to store new student data and its parent data
This commit adds functionality to store new student data and its parent data. The
student data is stored in the students table, while the parent data is stored in
the student_families table. The student data includes the student's name, gender,
NIK, place of birth, date of birth, address, province, city, district, village,
child number, siblings, and student image. The parent data includes the father's
name, NIK, phone number, education, job, income, and image, as well as the
mother's name, NIK, phone number, education, job, income, and image. The
functionality also includes compressing and storing the student and parent
images. If the data is successfully stored, a success message is displayed. If
not, an error message is displayed.